% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{cov_mmrr}
\alias{cov_mmrr}
\title{Calculate marginal covariance matrix of responses given predictors}
\usage{
cov_mmrr(X, Beta, Sigma, psi, type, num_nodes = 10)
}
\arguments{
\item{X}{An nr x p matrix of predictors.}

\item{Beta}{A p-vector of latent regression coefficients}

\item{Sigma}{An r x r covariance matrix for latent vector}

\item{psi}{An r-vector of conditional variance parameters}

\item{type}{An r-vector indicating response types:
1 means normal, 2 means Bernoulli,
and 3 means Poisson.}

\item{num_nodes}{Number of nodes for Gaussian quadrature used to
calculate predictions}
}
\value{
An nr x nr covariance matrix for responses given predictors (Y|X),
  in sparse format if n > 1.
}
\description{
Calculate marginal covariance matrix of responses given predictors
}
