% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_mmrr}
\alias{generate_mmrr}
\title{Generate data from mixed-type multivariate response regression model}
\usage{
generate_mmrr(X, Beta, R, type, psi)
}
\arguments{
\item{X}{An nr x p matrix of predictors. The first r elements are predictors
for the r responses in the first independent response vector, and so on.}

\item{Beta}{A p-vector of latent regression coefficients}

\item{R}{An r x r square root of the covariance matrix of the latent vectors.
Should be such that t(R) \%*\% R = solve(Omega).}

\item{type}{An r-vector indicating response types: 1 means normal, 2 means
Bernoulli, and 3 means Poisson.}

\item{psi}{Vector of variance parameters, ignored for Bernoulli
responses. For Poisson, generates Y| W as psi * Poi(exp(W) / psi)}
}
\value{
An n x r matrix of responses.
}
\description{
Generate data from mixed-type multivariate response regression model
}
