% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{lrt_approx}
\alias{lrt_approx}
\title{Approximate likelihood ratio testing for Latent Variables Multivariate
   Mixed-type Response Regression}
\usage{
lrt_approx(fit_null, fit_full, df, ...)
}
\arguments{
\item{fit_null}{The output from lvmmr() fit under the null hypothesis}

\item{fit_full}{The output from lvmmr() fit under the alternative hypothesis}

\item{df}{The degrees of freedom; if not supplied, set to the sum of the
difference in number of predictors and number of restrictions on Sigma.}

\item{...}{Other arguments passed to lvmmr() when re-fitting the full model
holding W fixed at fit_null$W.}
}
\value{
A list with the test statistic, p-value, and degrees of freedom.
}
\description{
Approximate likelihood ratio testing for Latent Variables Multivariate
   Mixed-type Response Regression
}
