% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{lvmmr}
\alias{lvmmr}
\title{Fit Latent Variables Mixed-type Multivariate Regression by PQL}
\usage{
lvmmr(
  Y,
  X,
  type,
  psi = rep(1, ncol(Y)),
  M = matrix(NA, nrow = ncol(Y), ncol = ncol(Y)),
  tol = rep(1e-08, 4),
  maxit = rep(100, 4),
  quiet = rep(TRUE, 4),
  relative = TRUE,
  pgd = FALSE,
  eps = 0,
  uni_fit = FALSE,
  Beta,
  Sigma,
  W,
  w_pen
)
}
\arguments{
\item{Y}{An n x r matrix of responses.}

\item{X}{An nr x p matrix of predictors or a list of length r whose ith
element is an n x p_i design matrix for the ith response.}

\item{type}{An r-vector indicating response types: 1 means Normal, 2 means
Bernoulli, and 3 means (quasi-)Poisson.}

\item{psi}{An r-vector of conditional variance parameters.}

\item{M}{An r x r matrix with restrictions for Sigma, with NA for unrestricted.}

\item{tol}{A 4-vector with tolerances for termination of: [1] overall
algorithm, [2] update of Beta and Sigma with W fixed, [3] update
of Sigma, and [4] update of W.}

\item{maxit}{A 4-vector with maximum number of iterations for the same steps
as the tol vector.}

\item{quiet}{A 4-vector indicating whether to print information for the
same steps as the tol vector.}

\item{relative}{If TRUE, use relative decrease of parameters to determine
convergence, otherwise use absolute.}

\item{pgd}{If TRUE, use projected gradient descent; ensures SPSD Sigma.}

\item{eps}{Lower bound for the smallest eigenvalue of Sigma, only used if
pgd = TRUE.}

\item{uni_fit}{If TRUE, fit r separate models. This requires (i) X is
a list or (ii) X is a matrix and r is a divisor of p. If (ii), it is
assumed that the first p / r columns of X correspond to the first
response, and so on.}

\item{Beta}{Initial iterate of regression coefficient vector. Either
a p-vector or a list of length r, where each element is the
coefficient vector for the ith response. Is obtained by fitting separate
GLMs if not supplied.}

\item{Sigma}{An r x r initial iterate for the latent covariance matrix.
Is set to diag(1e-3, ncol(Y)) if not supplied.}

\item{W}{An n x r initial iterate for the expansion points.
Is set to matrix(X %*% Beta, nrow = n, ncol = r, byrow = TRUE) if not
supplied.}

\item{w_pen}{Ridge penalty in W update; often useful to avoid overflows.
Defaults to largest eigenvalue of current Sigma iterate if not supplied.}
}
\value{
A list of final iterates and other information about the fit.
}
\description{
Fit Latent Variables Mixed-type Multivariate Regression by PQL
}
