% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR-package.R, R/MMRR.R
\docType{package}
\name{MMRR}
\alias{MMRR}
\title{MMRR: A package for performing multiple matrix regression with randomization}
\usage{
MMRR(Y, X, nperm = 999, scale = TRUE)
}
\arguments{
\item{Y}{The dependent variable in the form of a distance matrix}

\item{X}{A list of independent variables in the form of distance matrices (with optional names)}

\item{nperm}{The number of permutations to perform for significance testing}

\item{scale}{If TRUE, all variables are scaled}
}
\description{
The MMRR package provides functions for performing multiple matrix regression analysis
using randomization to calculate statistic significance.  These tests are also known
as multiple Mantel tests.  The package also provides a convenience function for plotting
the fitted MMRR models.

MMRR performs Multiple Matrix Regression with Randomization analysis
}
\examples{
Xmats <- list(geography = simData$geoMat, ecology = simData$ecoMat)
MMRR(simData$genMat, Xmats, nperm = 99)
}
