% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\docType{data}
\name{simData}
\alias{simData}
\title{simData}
\format{
A list of distance matrices
}
\usage{
data(simData)
}
\description{
A dataset containing three distance matrices for demonstrating example code
}
\examples{
data(simData)
}
\keyword{datasets}
