% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold.R
\name{unfold}
\alias{unfold}
\title{Unfold matrix}
\usage{
unfold(X, scale = TRUE)
}
\arguments{
\item{X}{A matrix}

\item{scale}{If TRUE, the resulting vector is scaled and centered}
}
\description{
Converts the lower diagonal elements of a matrix into a vector
}
