% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_beta}
\alias{update_beta}
\title{Update function for regression coefficients}
\usage{
update_beta(Y, X, W, Sigma, psi, type)
}
\arguments{
\item{Y}{Matrix (n x r) of responses.}

\item{X}{Matrix (nr x p) of predictors.}

\item{W}{Matrix (n x r) of expansion points (~predicted latent variables).}

\item{Sigma}{Covariance matrix(r x r) of the latent vector.}

\item{psi}{Vector (r x 1) of conditional variance parameters.}

\item{type}{Vector (r x 1) indicating response types.}
}
\value{
Updated vector of regression coefficients.
}
\description{
Update function for regression coefficients
}
