% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_Sigma_pgd}
\alias{update_Sigma_pgd}
\title{Accelerated projected gradient descent update function for Sigma}
\usage{
update_Sigma_pgd(
  R,
  D2,
  psi,
  Sigma.init,
  M,
  epsilon = 0,
  tol.dykstra = 1e-12,
  tol.ipiano = 1e-10,
  max.iter.dykstra = 1000,
  max.iter.ipiano = 1000,
  quiet = TRUE
)
}
\arguments{
\item{R}{Matrix (n x r) of working residuals.}

\item{D2}{Matrix (n x p) second derivatives of cumulant functions evaluated
at the expansion points.}

\item{psi}{Vector (r x 1) of conditional variance parameters.}

\item{Sigma.init}{Matrix (r x r ) with starting value for Sigma.}

\item{M}{Matrix (r x r) with restrictions for elements o Sigma
(NA = no restriction).}

\item{epsilon}{Scalar with lower bound on eigenvalues of Sigma.}

\item{tol.dykstra}{Scalar tolerance parameter for projection algorithm,}

\item{tol.ipiano}{Scalar tolerance for descent algorithm.}

\item{max.iter.dykstra}{Integer maximum number of iterations for projection
algorithm.}

\item{max.iter.ipiano}{Integer maximum number of iterations for descent
algorithm.}

\item{quiet}{Logical indicating whether to print information while running.}
}
\value{
Matrix (r x r) with update of Sigma
}
\description{
Accelerated projected gradient descent update function for Sigma
}
