% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_Sigma_trust}
\alias{update_Sigma_trust}
\title{Trust region update function for Sigma}
\usage{
update_Sigma_trust(
  Sigma_start,
  R,
  D2,
  psi,
  M,
  use_idx,
  tol = sqrt(.Machine$double.eps),
  maxit = 100
)
}
\arguments{
\item{Sigma_start}{Matrix (r x r ) with starting value for Sigma.}

\item{R}{Matrix (n x r) of working residuals.}

\item{D2}{Matrix (n x p) second derivatives of cumulant functions evaluated
at the expansion points.}

\item{psi}{Vector (r x 1) of conditional variance parameters.}

\item{M}{Matrix (r x r) with restrictions for elements o Sigma
(NA = no restriction).}

\item{use_idx}{Vector with indexes for observations to use in update.}

\item{tol}{Scalar tolerance parameter for trust region algorithm.}

\item{maxit}{Integer maximum number of iterations for trust region algorithm.}
}
\value{
Matrix (r x r) with update of Sigma
}
\description{
Trust region update function for Sigma
}
