% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{update_W}
\alias{update_W}
\title{Update function for expansion points (~predicted latent variables)}
\usage{
update_W(
  Y,
  X,
  W,
  Beta,
  Sigma,
  psi,
  type,
  pen = 1e-04,
  tol = 1e-08,
  maxit = 100,
  quiet = T
)
}
\arguments{
\item{Y}{Matrix (n x r) of responses.}

\item{X}{Matrix (nr x p) of predictors.}

\item{W}{Matrix (n x r) of expansion points (~predicted latent variables).}

\item{Beta}{Vector (p by 1) of regression coefficients.}

\item{Sigma}{Covariance matrix(r x r) of the latent vector.}

\item{psi}{Vector (r x 1) of conditional variance parameters.}

\item{type}{Vector (r x 1) indicating response types.}

\item{pen}{Regularization parameter lower bouding the eigenvalue of a
covariance matrix and multiplying an L2 penalty (see paper).}

\item{tol}{Tolerance parameter for trust region algorithm}

\item{maxit}{Maximum number of iterations for trust region algorithm.}

\item{quiet}{Logical indicating whether to print information while running.}
}
\value{
Matrix of updated expansion points
}
\description{
Update function for expansion points (~predicted latent variables)
}
