% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRR.R
\name{MMRR}
\alias{MMRR}
\title{MMMR}
\usage{
MMRR(Y, X, nperm = 999, scale = TRUE)
}
\arguments{
\item{Y}{The dependent variable in the form of a distance matrix}

\item{X}{A list of independent variables in the form of distance matrices (with optional names)}

\item{nperm}{The number of permutations to perform for significance testing}

\item{scale}{If TRUE, all variables are scaled}
}
\description{
MMRR performs Multiple Matrix Regression with Randomization analysis
}
