% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMMRR.R
\name{plotMMRR}
\alias{plotMMRR}
\title{Plot MMMR}
\usage{
plotMMRR(reg, Y, X, scale = TRUE, lineCol = "blue")
}
\arguments{
\item{reg}{The fitted MMRR model}

\item{Y}{The dependent variable in the form of a distance matrix}

\item{X}{A list of independent variables in the form of distance matrices (with optional names)}

\item{scale}{If TRUE, all variables are scaled}

\item{lineCol}{Color for regression line}
}
\description{
Plots the results of an MMRR analysis
}
\details{
The objects supplied for Y and X should be the same variables used to fit the MMRR model.  The paramter 'scale' should be the same as used to fit the model.
}
